requirements.txt
# VérificateurIA - Dépendances Python

# Interface Web
streamlit>=1.28.0
plotly>=5.17.0

# IA et LangChain
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.0.0
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0

# Recherche Web
tavily-python>=0.3.0

# Utilitaires
python-dotenv>=1.0.0

# Optionnel (pour développement)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
